% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertChar.R
\name{insertChar}
\alias{insertChar}
\title{Insert Special Character}
\usage{
insertChar(char_df = insertChar::w3charref)
}
\arguments{
\item{char_df}{dataframe of W3C character reference. See \code{?w3charref}.}
}
\value{
NULL value. Character codes are inserted when the "Insert *" button is
pressed
}
\description{
A Shiny Gadget to select special character and insert into the cursor
position. Insertion can be for HTML named entities or hexadecimal or decimal codes.
}
\examples{
if(interactive()){
  insertChar(char_df = insertChar::w3charref)
}
}
