% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertChar.R
\name{insertChar}
\alias{insertChar}
\title{Insert Special Character}
\usage{
insertChar(input_char_df = insertChar::w3charref)
}
\arguments{
\item{input_char_df}{dataframe of W3C character reference. See \code{?w3charref}.}
}
\value{
NULL value. Character codes are inserted when the "Insert *" button is
pressed
}
\description{
A Shiny Gadget to select special character and insert into the cursor
position. Insertion can be for HTML named entities or hexadecimal or decimal codes.
}
\details{
Other character sets not included in W3C may also be loaded if they are in the same
format as \code{\link{w3charref}}. Then the dataframe can be passed in to
\code{insertChar} as \code{insertChar(char_df = custom_charset)}.
}
\examples{
if(interactive()){
  insertChar(input_char_df = insertChar::w3charref)
}
}
